+++ src/Expr/BoolOpsExpr.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace AndrewGos\QueryBuilder\Expr;
 
+use AndrewGos\QueryBuilder\Exception\QueryBuilderException;
 use AndrewGos\QueryBuilder\Grammar\GrammarInterface;
 use AndrewGos\QueryBuilder\Helper\HExpr;
 use AndrewGos\QueryBuilder\Query\Select\SelectQueryInterface;
@@ -56,6 +57,10 @@ class BoolOpsExpr extends AbstractExpr
     protected function doBuild(GrammarInterface $grammar): array
     {
         $conditions = HExpr::normalizeConditions($this->conditions, $grammar);
+
+        if (empty($conditions)) {
+            throw QueryBuilderException::emptyBoolExpression();
+        }
 
         $expressions = [];
         $params = [];